Separates the S3 storage into its own file
